+++ lib/osm.rb
@@ -64,19 +64,6 @@ Dir[File.join(File.dirname(__FILE__) , 'osm', '*.rb')].each {|file| require file
 
 module Osm
 
-    # Configure the options used by classes in the module
-    # @param [Hash] options
-    # @option options [Hash] :cache_config (optional) How classes in the module will cache data. Whilst this is optional you should remember that caching is required to use the OSM API.
-    # @option options[:cache] [Class] :cache An instance of a cache class, must provide the methods (exist?, delete, write, read), for details see Rails.cache.
-    # @option options[:cache] [Fixnum] :ttl (optional, default = 30.minutes) The default TTL value for the cache, note that some items are cached for twice this time and others are cached for half this time (in seconds)
-    # @option options[:cache] [String] :prepend_to_key (optional, default = 'OSMAPI') Text to prepend to the key used to store data in the cache
-    # @return nil
-    def self.configure(options)
-      Osm::Model.configure(options[:cache])
-      nil
-    end
-
-
   private
   def self.make_datetime(date, time, options={})
     date = nil if date.nil? || date.empty? || (!options[:ignore_epoch] && epoch_date?(date))